first_linear_net.py trains

+++ training_utils.py
@@ -27,7 +27,7 @@ def train(loader_train, model, loss_fn, optimizer, dtype,
         model.train()
         for t, (x, y) in enumerate(loader_train):
             x_var = Variable(x.type(dtype))
-            y_var = Variable(y.type(dtype).long())
+            y_var = Variable(y.type(dtype))
 
             scores = model(x_var)
